+++ CHANGELOG.md
@@ -22,7 +22,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+*
+
+## [0.23.0] - 2022-06-03
+
+### Added
+
+* Add `input-file` component - [ripe-copper/#48](https://github.com/ripe-tech/ripe-copper/issues/48)
+* Add slots to `input-symbol` component - [ripe-util-vue/#259](https://github.com/ripe-tech/ripe-util-vue/issues/259)
+* Button icon dropdown global hide prop - [ripe-pulse/#281](https://github.com/ripe-tech/ripe-pulse/issues/281)
+* Add `avatar-list` component - [#571](https://github.com/ripe-tech/ripe-components-vue/issues/571)
+* Add node `16`, `17` and `18` support
+* Add `disabled` prop to `button-dropdown` - [ripe-util-vue/#288](https://github.com/ripe-tech/ripe-util-vue/issues/288)
+* Add `update:sorted-items` event in `table-ripe` - [ripe-util-vue/#288](https://github.com/ripe-tech/ripe-util-vue/issues/288)
+
+### Changed
+
+* Removed `announcements` component and it's related logic from `header-ripe` component
+
+### Fixed
+
 * Prop issue in `button-icon-dropdown` when propagating `disabled`
+* Fix `button-dropdown` events not being propagated correctly - [ripe-util-vue/#288](https://github.com/ripe-tech/ripe-util-vue/issues/288)
+* Fix storybook dependency problem after minor bump, lock versions to minor - [ripe-welcome/PR#187](https://github.com/ripe-tech/ripe-welcome/pull/187)
+* Fix eslint dependencies problems
+* Fix `code-editor` `value` emit event - [ripe-util-vue/#305](https://github.com/ripe-tech/ripe-util-vue/issues/305)
 
 ## [0.22.0] - 2022-04-22